feat: likeCount, commentCount를 entity 안에 넣고 피드 조회 Footprint 기준으로 변경

+++ src/main/kotlin/zip/cafe/api/dto/Feed.kt
@@ -40,9 +40,9 @@ data class FeedReview(
     val finalScore: Double,
     val images: List<FeedImage>,
     val keyword: List<FeedKeyword>,
-    val likeCount: Int,
+    val likeCount: Long,
     val content: String,
-    val commentCount: Int,
+    val commentCount: Long,
     @JsonFormat(pattern = defaultDateTimeFormat)
     val createdAt: LocalDateTime
 )
